+++ src/lang/compile-buildin.js
@@ -2,6 +2,7 @@ import { cmp, cmpAnalytical, cmpNatural, cmpNaturalAnalytical } from '../utils/c
 import { hasOwn, addToSet, getPropertyValue, isPlainObject, isRegExp, isArrayLike, isTruthy, parseIntDefault, isArray } from '../utils/misc.js';
 
 export default Object.freeze({
+    unsafeRef,
     ensureArray,
     bool: isTruthy,
     and: (a, b) => isTruthy(a) ? b : a,
@@ -35,6 +36,19 @@ export default Object.freeze({
     filter,
     slice
 });
+
+function unsafeRef(fn, name, range) {
+    try {
+        return fn();
+    } catch (e) {
+        if (name) {
+            throw Object.assign(
+                new ReferenceError(`Cannot access $${name} before initialization`),
+                { details: { name, loc: { range } } }
+            );
+        }
+    }
+}
 
 function ensureArray(value) {
     return isArray(value) ? value : [value];